requirements.txt
pandas>=2.2.0
groq>=0.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
tabulate>=0.9.0